tests/submission/Infraware/Session History/contents/Session_History/Not Use/history_go_over.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_go_over</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		var t1 = async_test('history go over test');
		var poped = false, initialURL = location.href;
		
		var len;
		var delta;
		len = window.history.length;
		delta = 10 + len;

		window.history.go(delta);

		function onPopstateEvent(e) {
			if (initialURL != location.href)
				poped = true;
		}

		if (window.addEventListener) {
            window.addEventListener('popstate', onPopstateEvent, false);
        }
        else if (window.attachEvent) {
            window.attachEvent("onpopstate", onPopstateEvent);
        }
        else {
			test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			});
        }
		
		setTimeout(function(){
			t1.step(function() {
				assert_false(poped, "popstate event is raised.");
			});
			t1.done();
		}, 500);

	</script>
 </body>
</html>
